refactor text_summarization application
