+++ applications/text_summarization/deploy/paddle_serving/pipeline_client.py
