+++ applications/text_summarization/deploy/paddle_serving/pipeline_service.py
