+++ examples/text_summarization/unimo-text/deploy/paddle_serving/pipeline_client.py
@@ -11,9 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from paddle_serving_server.pipeline import PipelineClient
 import time
+
 import numpy as np
+from paddle_serving_server.pipeline import PipelineClient
+
 from paddlenlp.utils.log import logger
